// app/src/generated/models/calendarFeedResponseFeedType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Scope of events included in the calendar feed
 */
export type CalendarFeedResponseFeedType =
  (typeof CalendarFeedResponseFeedType)[keyof typeof CalendarFeedResponseFeedType];

export const CalendarFeedResponseFeedType = {
  ALL_PAYMENTS: 'ALL_PAYMENTS',
  CONTRACT: 'CONTRACT',
  PROPERTY_PAYMENTS: 'PROPERTY_PAYMENTS',
  TENANT_PAYMENTS: 'TENANT_PAYMENTS',
} as const;