// app/src/generated/models/createCalendarFeedRequest.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCalendarFeedRequestFeedType } from './createCalendarFeedRequestFeedType';

export interface CreateCalendarFeedRequest {
  /** Scope of events included in the calendar feed */
  feedType: CreateCalendarFeedRequestFeedType;
  /** @nullable */
  contractIdentifier?: string | null;
  /** @nullable */
  propertyIdentifier?: string | null;
  /** @nullable */
  tenantIdentifier?: string | null;
}